Fixed: Some cases of unused functions
