+++ projects/dxt-lossless-transform-bc3/benches/transform_standard/portable32.rs
@@ -1,5 +1,7 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc3::transforms::standard::transform::bench_exports::*;
+use dxt_lossless_transform_bc3::transforms::standard::transform::bench::{
+    portable32::u32_unroll_2, *,
+};
 use safe_allocator_api::RawAlloc;
 
 fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
@@ -14,7 +16,7 @@ fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut R
 
 fn bench_portable32_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        u32_transform_unroll_2(
+        u32_unroll_2(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
