+++ projects/dxt-lossless-transform-bc3/src/transforms/standard/transform/mod.rs
@@ -207,24 +207,4 @@ pub unsafe fn split_blocks_with_separate_pointers(
 
 // Re-export functions for benchmarking when the 'bench' feature is enabled
 #[cfg(feature = "bench")]
-#[allow(clippy::missing_safety_doc)]
-pub mod bench_exports {
-    pub unsafe fn u32_transform(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
-        super::portable32::u32(input_ptr, output_ptr, len)
-    }
-
-    pub unsafe fn u32_transform_unroll_2(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
-        super::portable32::u32_unroll_2(input_ptr, output_ptr, len)
-    }
-
-    #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
-    pub unsafe fn u32_avx2_transform(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
-        super::avx2::u32_avx2(input_ptr, output_ptr, len)
-    }
-
-    #[cfg(feature = "nightly")]
-    #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
-    pub unsafe fn avx512_vbmi_transform(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
-        super::avx512::avx512_vbmi(input_ptr, output_ptr, len)
-    }
-}
+pub mod bench;
